🎨[REFACTOR] MenuListViewController 클래스, CustomCollectionViewCell 클래스 파일로 분리
